hw_utils/hwmquery: Fixed a compiler warning in 64 bit due to ssize_t.
git-svn-id: file:///svn/uhexen2/code/trunk@4306 eaf649e6-6c13-0410-b194-84221022858e

+++ hw_utils/hwmquery/hwmquery.c
@@ -321,7 +321,7 @@ int main (int argc, char **argv)
 			printf (" NONE\n");
 		else
 		{
-			printf (" %d entries\n", size / 6);
+			printf (" %d entries\n", (int)size / 6);
 			if (size % 6 != 0)
 				printf ("Warning: not counting truncated last entry\n");
 			// each address is 4 bytes (ip) + 2 bytes (port) == 6 bytes
